refactor: add notification service to create and handle notification request
