+++ actions/push.js
@@ -2,29 +2,7 @@ var program = require('commander');
 var request = require('request');
 
 var credentialLib = require('../libraries/credential');
-var requestHandler = require('../handlers/request');
-
-function createPushData (message, credential) {
-	var requestPushData = {
-		method: 'POST',
-		url: 'https://go.urbanairship.com/api/push/',
-		
-		json: {
-			'audience': 'all',
-			'notification': {'alert': message},
-			'device_types': 'all',
-		},
-
-		headers: {
-			'Authorization': 'Basic ' + credentialLib.encode(credential),
-			'Accept': 'application/vnd.urbanairship+json; version=3;',
-			'Content-Type': 'application/json',
-		},
-	}
-
-	return requestPushData;
-}
-
+var notification = require('../services/notification');
 
 function pushAction (message) {
 	var credential = null;
@@ -34,9 +12,9 @@ function pushAction (message) {
 
 	try {
 		credential = credentialLib.load(program);
-		requestPushData = createPushData(message, credential);
+		requestPushData = notification.createRequest(message, null, credential);
 
-		request(requestPushData, requestHandler);
+		request(requestPushData, notification.handleRequest);
 	} catch (error) {
 		console.log(error);
 	}
