--- handlers/request.js
@@ -1,7 +1,0 @@
-function requestHandler (error, request, body) {
-	if (error) return console.log('Error: ', error);
-
-	console.log('Notification sent:', body);
-}
-
-module.exports = requestHandler;
